project/requirements.txt
aiofiles==0.6.0
fastapi==0.65.2
Jinja2==2.11.3
pytest==6.2.4
redis==3.5.3
requests==2.25.1
uvicorn[standard]
alembic==1.7.1
asyncpg==0.24.0
sqlmodel
psycopg2-binary==2.9.1
aioredis
fastapi-cache2[redis]
gunicorn